Refactor main method to use companion object
