--- src/main/kotlin/com/hhandoko/luwak/resources/OrderResource.kt
@@ -1,36 +1,0 @@
-/**
- * File     : OrderResource.kt
- * License  :
- *   Copyright (c) 2017 Herdy Handoko
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *           http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-package com.hhandoko.luwak.resources
-
-import javax.ws.rs.GET
-import javax.ws.rs.Path
-import javax.ws.rs.Produces
-import javax.ws.rs.core.MediaType
-
-import com.hhandoko.luwak.api.Order
-
-@Path("orders")
-@Produces(MediaType.APPLICATION_JSON)
-class OrderResource {
-
-    @GET
-    fun get(): Order {
-        return Order("Coffee")
-    }
-
-}
